Release 0.2.3
This commit fix the branch creation process. Before this commit, creating a branch from another one
raised an error if the original branch was not found in the remote. And this is because before creating
a new branch gith updates the original branch, doing 'git pull origin current_branch'.

This commit adds a try-except to the pull process. Now, if the branch does not exist in remote, it will be
catched, and gith do not block the branch creation. It just print an Info message in the shell and creates
the branch.

This is a common case when someone creates a branch, local branch, and do not push anything, but still wants
to create a new branch from that one.